Introduce non-generic IStructId for string-backed ids
Strings are not structs so we cannot use them in conjunction with the type constraints in IStructId<T>. It simplifies things to have a non-generic interface for the string-version of a struct id, so we can keep the type constraints very specific for structs and other primitives.

The analyzer project is not renamed to StructId.Analyzer so we can instead have a StructId project for development of the main interfaces and code so it can be tested and compiled independently from its inclusion in the test project (which still simulates consumption scenario more closely).