Add support for IParsable<T>
Ths is a requirement for using struct ids in ASP.NET core endpoints, for example.

This is implemented such that if in a given `IStructId<T>`,  the `T : IParsable<T>`, the parsable implementation is provided as a pass-through to the value implementation.

For `IStructId` which is for string-typed values, we always emit a simple implementation that just create a new struct id.