Introduce StructId.New(value) via static interface, Json.NET support
The factory method `New(Value)` makes it easier to implement the Json.NET converters since then we don't need to go through the IParsable interface like we do for STJ. Not sure if there's a gain to be had on that front too.

Moved integration/functional tests to another project, which just consumes the analyzer/generator at compile-time, rather than for tests itself.