Rename placeholder type to TSelf, Self
The more common template type will be TSelf, which is an IStructId<TValue>. We can therefore simply using TSelf and TValue across the templates which is more intuitive.

The Self (without T) is for the string-version only, which may or may not be required depending on the concrete scenario being generated.