Initial user templates support
Templates are regular C# code, annotated with `[StructId]`.

The template must provide the `[TId] Value` primary constructor, so that operations for the provided members can use the value just as it will for generated `IStructId` implementations. The type of the value determines the target ids that will get the template applied: they must match the TId type.

The provided functional sample showcases how it works. Say you have a custom interface in your app, `IId` which you use to retrieve a GUID that represents the ID of your domain objects. You can have all struct ids that have a TId of Guid implement this interface by defining the template as:

```csharp
[TStructId]
file partial record struct IIdTemplate(Guid Value) : IId
{
    public Guid Id => Value;
}
```

What happens at compile-time:
1. The attribute triggers template processing for this type
2. The `Guid Value` parameter is used to filter struct ids that implement `IStructId<Guid>` (which will therefore have a matching constructor and value type).
3. The emitted partial is added to whichever namespace the target ID is, so the template does not need a namespace at all (should use global namespace).
4. The `file` keyword is used so that this type does not pollute the API surface of the project containing the template.

Templates can come from referenced projects too, as long as the generator can locate the original syntax tree for it (to apply it to the target id).