Generalize value templates too, automatically adding Ulid to Dapper
A new templatized value feature allows built-in compiled template support for Ulid as a value type for struct ids. The template follows the same pattern as the struct id template (pending analyzer/codefix) and the dapper generator has been extended to consider them when emitting the registration code.

We keep the original custom handler support too, which would still win over templatized one if users provide a type handler in the compilation themselves (as shown in the functional tests via the `StringUlidHandler` type.