Add built-in support for parsable/formattable values for EF
Rather than adding explicit Ulid support as suggested in #18, a more flexible approach is to simply rely on the combination of IParsable<T> plus IFormattable checks on the TValue/TId and emit a standard EF value converter that relies on those, plus the INewable<TSelf> we already have for the struct-id itself.

This involved a more flexible resource template selection mechanism, which needs further refactoring for simplification (and perhaps dropping altogether) of the BaseGenerator.