Make feature-dependent templates conditional
As we moved to compiled templates for EF/Dapper/Ulid, we need to only include those templates in the compilation if the user actually references the respective packages.

For now, this is not an extensible mechanism nad we just detect the dependency on the relevant packages via MSBuild and remove the templates as needed.

We moved the Ulid support for New() (mirroring the support we offer for TId.New() for Guids) since it can now be conditional too.